worker/igdownloader/requirements.txt
instaloader==4.7.1
python-dotenv==0.17.1
threadpool==1.3.2